Skeleton and initial functionality
This is going to be an app to list and track what items you have packed for your travels. It is quite simple in functionality but i feel touches on a lot of essential react from passing state as porps and other forms of state management to controlled components and more. I have made the app skeleton and will build the functionality hopefully over the next couple of days