Application Finished
I added the rest of the logic for this application and hvae synced it all with the UI using state and state management. I have given features such as sorting clearing and checking when an items is packed as well as statistic trackers for items packed and percentage of total packed using the sort method on shallow copies of the items state before rendering. The application is mostly done, i will add some color changes soon to make it look a little less bleak.